+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Fixed a bug where fields with hidden labels would refer to themselves as `__blank__` in validation errors. ([#6699](https://github.com/craftcms/cms/issues/6699))
 - Fixed a bug where it was impossible to use lists in directive arguments. ([#6693](https://github.com/craftcms/cms/issues/6693))
 - Fixed a bug where asset indexing utility would not work correctly if no files were found on the selected volumes. ([#6658](https://github.com/craftcms/cms/issues/6658))
+- Fixed a bug where project config files could get deleted if the casing of a component’s handle was changed, on case-insensitive file systems. ([#6708](https://github.com/craftcms/cms/issues/6708))
 
 ## 3.5.5 - 2020-08-17